Rename mod.rs to mm2_tests.rs
It's better if the file names are descriptive, helps navigating the files in the IDE.
`mm2_tests.rs` hints that we're working with the `mm2_tests` Rust module, whereas `mod.rs` hides obscures this.
Likewise, putting files *deeper* in the file hierarchy makes them harder to find. Consequentially it's often better to grow the hierarchy organically, e.g. adding levels of hierarchy when we have a surplus of similar files.